Move error formatting out of the storage component
